requirements.txt
Flask==3.1.*
pdfplumber==0.11.*
fpdf2==2.8.*
gunicorn==23.0.*
